Create review controller

+++ BE_Book_Food/src/main/java/com/example/ecommerce/book_food/dto/respone/ReviewResponse.java
@@ -1,5 +1,6 @@
 package com.example.ecommerce.book_food.dto.respone;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -12,7 +13,7 @@ import lombok.NoArgsConstructor;
 public class ReviewResponse {
     private Long id;
     private String username;
-    private Integer rating; // 1 - 5
+    private BigDecimal rating; // 1 - 5
     private String comment;
     private LocalDateTime createdAt;
 }
